separate access tests from validation tests
